# src/swagger.json
{
	"openapi": "3.0.1",
	"info": {
		"title": "REST API for SmartSense Gaia-X Signer Tool",
		"version": "1.0.0"
	},
	"servers": [{ "url": "http://localhost:8000/" }],
	"paths": {
		"/createWebDID": {
			"post": {
				"summary": "Create Web DID",
				"description": "Required domain name to create did:web",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateDIDSchemaBody"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessSchemaResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorSchemaResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorSchemaResponse"
								}
							}
						}
					}
				}
			}
		},
		"/onBoardToGaiaX": {
			"post": {
				"summary": "On board to Gaia-X",
				"description": "Generate Legal Person and Service Offer Credentials",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OnBoardGaiaXSchemaBody"
							},
							"examples": {
								"LegalParticipant": {
									"value": {
										"domain": "greenworld.proofsense.in",
										"templateId": "LegalParticipant",
										"privateKeyUrl": "https://example.com",
										"data": {
											"legalName": "Green World",
											"legalRegistrationType": "leiCode",
											"legalRegistrationNumber": "9695007586GCAKPYJ703",
											"headquarterAddress": "BE-BRU",
											"legalAddress": "BE-BRU"
										}
									}
								},
								"ServiceOffering": {
									"value": {
										"domain": "dev.smartproof.in",
										"tenant": "smart",
										"templateId": "ServiceOffering",
										"privateKeyUrl": "https://example.com",
										"data": {
											"name": "AWS S3 Service",
											"fileName": "s3Service.json",
											"description": "Amazon S3 bucket is a public cloud storage resource.",
											"policyUrl": "https://aws.amazon.com/privacy/?nc1=f_pr",
											"termsAndConditionsUrl": "https://aws.amazon.com/service-terms/",
											"termsAndConditionsHash": "4f598faf41fb29a44d63ec413730d55fabd2d249f1976fee1c8eec8003d7f539",
											"requestType": "API",
											"accessType": "digital",
											"formatType": "application/json"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessSchemaResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorSchemaResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/createVC": {
			"post": {
				"summary": "Create verifiable credential",
				"description": "Generate verifiable credential along with proof",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateVCSchemaBody"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessSchemaResponse"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorSchemaResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorSchemaResponse"
								}
							}
						}
					}
				}
			}
		},
...
		"/get/trust-index": {
			"post": {
				"summary": "Calculation of Trust index, Veracity & Transperency",
				"description": "Calculates Trust index, Veracity & Transperency using holder VC and self description of service offering",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GetTrustIndexSchemaBody"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Success",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessSchemaResponse"
								},
								"examples": {
									"trustIndex": {
										"summary": "Trust index, Veracity & Transperency of service offering",
										"value": {
											"message": "Success",
											"data": {
												"veracity": 0.25,
												"transparency": 1.4,
												"trustIndex": 0.825,
												"certificateDetails": {
													"validFrom": "Apr 21 00:00:00 2023 GMT",
													"validTo": "Apr 20 23:59:59 2024 GMT",
													"subject": {
														"jurisdictionCountry": "DE",
														"jurisdictionSate": "Hamburg",
														"jurisdictionLocality": "Hamburg",
														"businessCategory": "Private Organization",
														"serialNumber": "HRB 170364",
														"country": "DE",
														"state": "Hamburg",
														"locality": "Hamburg",
														"organization": "deltaDAO AG",
														"commonName": "www.delta-dao.com"
													},
													"issuer": {
														"commonName": "DigiCert EV RSA CA G2",
														"organization": "DigiCert Inc",
														"country": "US"
													}
												}
											}
										}
									}
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorSchemaResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorSchemaResponse"
								}
							}
						}
					}
				}
			}
		},
...
	},
	"components": {
		"schemas": {
			"SuccessSchemaResponse": {
				"properties": {
					"data": {
						"type": "object"
					},
					"message": {
						"type": "string"
					}
				}
			},
			"ErrorSchemaResponse": {
				"properties": {
					"error": {
						"type": "string"
					},
					"message": {
						"type": "string"
					}
				}
			},
			"CreateDIDSchemaBody": {
				"required": ["domain"],
				"properties": {
					"domain": {
						"type": "string",
						"example": "dev.smartproof.in"
					},
					"tenant": {
						"type": "string",
						"example": "smart"
					},
					"services": {
						"type": "array",
						"items": {
							"type": "object",
							"properties": {
								"type": {
									"type": "string",
									"example": "CredentialRegistry"
								},
								"serviceEndpoint": {
									"type": "string",
									"example": "https://ssi.eecc.de/api/registry/vcs"
								}
							}
						}
					}
				}
			},
			"OnBoardGaiaXSchemaBody": {
				"required": ["domain"],
				"properties": {
					"domain": {
						"type": "string",
						"example": "gaiaxapi.proofsense.in"
					},
					"tenant": {
						"type": "string",
						"example": "Hella"
					},
					"templateId": {
						"type": "string",
						"example": "LegalParticipant"
					},
					"privateKeyUrl": {
						"type": "string",
						"example": "https://example.com"
					},
					"data": {
						"type": "object"
					}
				}
			},
			"CreateVCSchemaBody": {
				"required": ["issuerDid", "subjectDid", "templateId", "privateKeyUrl", "credentialOffer"],
				"properties": {
					"issuerDid": {
						"type": "string",
						"example": "did:web:gaiaxapi.proofsense.in:Hella"
					},
					"subjectDid": {
						"type": "string",
						"example": "did:web:gaiaxapi.proofsense.in:Hella"
					},

					"templateId": {
						"type": "string",
						"example": "LegalParticipant"
					},
					"privateKeyUrl": {
						"type": "string",
						"example": "https://example.com"
					},

					"credentialOffer": {
						"type": "object",
						"properties": {
							"legalName": {
								"type": "string",
								"example": "Hella"
							},
							"legalRegistrationType": {
								"type": "string",
								"example": "vatID"
							},
							"legalRegistrationNumber": {
								"type": "string",
								"example": "FR79537407926"
							},
							"headquarterAddress": {
								"type": "string",
								"example": "BE-BRU"
							},
							"legalAddress": {
								"type": "string",
								"example": "BE-BRU"
							}
						}
					}
				}
			},
...
			"GetTrustIndexSchemaBody": {
				"required": ["participantSD", "serviceOfferingSD"],
				"properties": {
					"participantSD": {
						"type": "string",
						"example": "https://www.delta-dao.com/.well-known/participant.json"
					},
					"serviceOfferingSD": {
						"type": "string",
						"example": "https://delta-dao.com/.well-known/serviceGaiaxDLTNetwork.json"
					}
				}
			},
...
		}
	}
}